Crash when setting 'alt' CSS property to inherit or initial https://bugs.webkit.org/show_bug.cgi?id=138774 <rdar://problem/18995409>
Reviewed by Chris Fleizach.

Source/WebCore:

The StyleResolver was not handling 'inherit' or 'initial' values for
'alt' CSS property. As a result, the code was assuming the CSSValue
was a CSSPrimitiveValue (although it could be a CSSInitialValue or
a CSSInheritValue). In such case, the code would dereference a null
pointer because primitiveValue is initialized using:
CSSPrimitiveValue* primitiveValue = is<CSSPrimitiveValue>(*value)
    ? downcast<CSSPrimitiveValue>(value)
    : nullptr;

Test: fast/css/alt-inherit-initial.html

* css/StyleResolver.cpp:
(WebCore::StyleResolver::applyProperty):

LayoutTests:

Add a layout test to verify that setting the 'alt' CSS property to
inherit / initial is not crashing and working as intended.

* fast/css/alt-inherit-initial-expected.txt: Added.
* fast/css/alt-inherit-initial.html: Added.


Identifier: 156619@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@176161 268f45cc-cd09-0410-ab3c-d52691b4dbfc