AX: com.apple.WebKit.WebContent crashed at WebCore: WebCore::AccessibilityRenderObject::computeAccessibilityIsIgnored const https://bugs.webkit.org/show_bug.cgi?id=138905
Reviewed by Mario Sanchez Prada.

The crash log indicates that m_renderer is null at the time we ask ancestorsOfType().
This is more of a speculative fix, since I am not entirely sure m_renderer is null when we enter the method.

Unable to determine cause of crash or how to reproduce on demand.

* accessibility/AccessibilityRenderObject.cpp:
(WebCore::AccessibilityRenderObject::computeAccessibilityIsIgnored):


Identifier: 156864@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@176484 268f45cc-cd09-0410-ab3c-d52691b4dbfc