AX: [ATK] Unskip the skipped ATK tests in roles-exposed.html https://bugs.webkit.org/show_bug.cgi?id=139016
Reviewed by Chris Fleizach.

Source/WebCore:

There was one test that could not be unskipped without making a change
for that one to pass: mapping DocumentRegionRole to ATK_ROLE_PANEL. The
rest of the "fix" is just unskipping tests for ATK and updating the test
expectations.

No new tests needed.

* accessibility/atk/WebKitAccessibleWrapperAtk.cpp:
(atkRole):

LayoutTests:

* accessibility/roles-exposed.html: Unskipped a bunch of tests.
* platform/efl/accessibility/roles-exposed-expected.txt: Updated for unskipped tests.
* platform/gtk/accessibility/roles-exposed-expected.txt: Updated for unskipped tests.


Identifier: 156895@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@176515 268f45cc-cd09-0410-ab3c-d52691b4dbfc