AX: [ATK] Meter and Option elements do not expose their id attribute https://bugs.webkit.org/show_bug.cgi?id=139017
Reviewed by Mario Sanchez Prada.

Source/WebCore:

The options in a collapsed select element lack a node, so get the id
attribute from the associated action element. In the case of a meter,
the meter element itself is not exposed; its RenderMeter is instead.
So associate the meter element's id with the exposed RenderMeter.

No new tests. Instead, updated existing expectations to reflect the fix.

* accessibility/AccessibilityObject.h:
* accessibility/AccessibilityProgressIndicator.cpp:
(WebCore::AccessibilityProgressIndicator::element):
* accessibility/AccessibilityProgressIndicator.h:
* accessibility/atk/WebKitAccessibleWrapperAtk.cpp:
(webkitAccessibleGetAttributes):

LayoutTests:

* accessibility/roles-exposed.html: Unskipped the meter element test.
* platform/efl/accessibility/roles-exposed-expected.txt: Updated to reflect fixes.
* platform/gtk/accessibility/roles-exposed-expected.txt: Updated to reflect fixes.


Identifier: 156931@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@176552 268f45cc-cd09-0410-ab3c-d52691b4dbfc