Change CallFrame::globalThisValue() to not use CallFrame::scope() https://bugs.webkit.org/show_bug.cgi?id=139202
Reviewed by Mark Lam.

Changed to use the globalThis() on the globalObject associated with the
callee.  Moved the inline definition to JSGlobalObject.h instead of
including JSGlobalObject.h in JSScope.h.  Also moved it as JSScope
objects are no longer involved in getting the value.

* runtime/JSGlobalObject.h:
(JSC::ExecState::globalThisValue):
* runtime/JSScope.h:
(JSC::ExecState::globalThisValue): Deleted.


Identifier: 157048@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@176700 268f45cc-cd09-0410-ab3c-d52691b4dbfc