If the action menu hit test times out, a menu with a single separator appears https://bugs.webkit.org/show_bug.cgi?id=139320 <rdar://problem/19158331>
Reviewed by Beth Dakin.

* UIProcess/mac/WKActionMenuController.h:
* UIProcess/mac/WKActionMenuController.mm:
(-[WKActionMenuController menuNeedsUpdate:]):
(-[WKActionMenuController _defaultMenuItems]):
Add a new "TimedOut" state, which we get into if the sync wait for
didPerformActionMenuHitTest times out. In this case, we'll drop the
separator item and give up on the menu. This way, the menu will never
have just a single separator item at the end of menuNeedsUpdate:.


Identifier: 157176@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@176883 268f45cc-cd09-0410-ab3c-d52691b4dbfc